rm master data error list
